feat: added tab button icon
