+++ config.json
@@ -76,6 +76,7 @@
     {
       "action": "general.playersList",
       "actionHold": "",
+      "icon": "",
       "label": "TAB"
     }
   ]
