+++ src/controls.ts
@@ -927,3 +927,14 @@ export function updateBinds (commands: any) {
     }))
   }
 }
+
+export const onF3LongPress = async () => {
+  const select = await showOptionsModal('', f3Keybinds.filter(f3Keybind => {
+    return f3Keybind.mobileTitle && (f3Keybind.enabled?.() ?? true)
+  }).map(f3Keybind => {
+    return `${f3Keybind.mobileTitle}${f3Keybind.key ? ` (F3+${f3Keybind.key})` : ''}`
+  }))
+  if (!select) return
+  const f3Keybind = f3Keybinds.find(f3Keybind => f3Keybind.mobileTitle === select)
+  if (f3Keybind) void f3Keybind.action()
+}
