+++ src/react/MobileTopButtons.module.css
@@ -35,3 +35,12 @@
   padding: 4px 6px;
   outline: 0.5px solid white;
 }
+
+.tab-btn {
+  color: #fff;
+  background: #9c8c86;
+  outline: .5px solid #fff;
+  padding: 3px 5px;
+  font-family: pixelarticons, mojangles, monospace;
+  font-size: 12px;
+}
